updated error logging

+++ src/events/guildDelete.js
@@ -1,10 +1,12 @@
 module.exports = {
     event: "guildDelete",
     run: async (client, db, guild) => {
-        const guilds = db.collection('Guilds');
+        if (guild.available && client.isReady()) {
+            const guilds = db.collection('Guilds');
 
-        guilds.deleteOne({ guildID: guild.id });
+            guilds.deleteOne({ guildID: guild.id });
 
-        console.log(`LEFT GUILD: ${guild.name} (${guild.id})`);
+            console.log(`LEFT GUILD: ${guild.name} (${guild.id})`);
+        }
     }
 };
